Creating an EditExpAutoSubmitPage page, adding a ScheduleTab widget, refactoring the code, adding localization for the ServiceList class

+++ packages/core_repository/lib/history_repository.dart
@@ -1,5 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:core_data/core_data.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/foundation.dart';
 
 class HistoryRepository {
@@ -13,9 +14,13 @@ class HistoryRepository {
     required String docNumber,
     required List<Map<String, dynamic>> fines,
   }) async {
+    final user = FirebaseAuth.instance.currentUser;
+    if (user == null) throw Exception("User is not signed in");
+
     final docRef = firestore.collection('fines_history').doc();
 
     final data = {
+      'userId': user.uid, // важно!
       'carNumber': carNumber.trim().toUpperCase(),
       'docSeries': docSeries,
       'docNumber': docNumber,
@@ -24,14 +29,19 @@ class HistoryRepository {
     };
 
     await docRef.set(data);
+
     if (kDebugMode) {
       print('✅ History added: ${docRef.id}');
     }
   }
 
   Stream<List<FineHistory>> getHistory(String carNumber) {
+    final user = FirebaseAuth.instance.currentUser;
+    if (user == null) throw Exception("User is not signed in");
+
     return firestore
         .collection('fines_history')
+        .where('userId', isEqualTo: user.uid)
         .where('carNumber', isEqualTo: carNumber.trim().toUpperCase())
         .orderBy('checkedAt', descending: true)
         .snapshots()
@@ -39,12 +49,19 @@ class HistoryRepository {
   }
 
   Future<void> deleteAll(String carNumber) async {
-    final snapshot =
-        await firestore.collection('fines_history').where('carNumber', isEqualTo: carNumber.trim().toUpperCase()).get();
+    final user = FirebaseAuth.instance.currentUser;
+    if (user == null) throw Exception("User is not signed in");
+
+    final snapshot = await firestore
+        .collection('fines_history')
+        .where('userId', isEqualTo: user.uid)
+        .where('carNumber', isEqualTo: carNumber.trim().toUpperCase())
+        .get();
 
     for (var doc in snapshot.docs) {
       await doc.reference.delete();
     }
+
     if (kDebugMode) {
       print('✅ All history removed for $carNumber');
     }
@@ -53,6 +70,7 @@ class HistoryRepository {
   Future<void> deleteSingle(String docId) async {
     final docRef = firestore.collection('fines_history').doc(docId);
     await docRef.delete();
+
     if (kDebugMode) {
       print('✅ Record $docId removed');
     }
